Update 09-vascular-trauma.Rmd
completed cerebrovascular trauma and endovascular trauma sections. zotero added references as indicated in the in-line text with links to resources. if the formatting with spacing of the bullet points is incorrect let me know, had to re-do it as it did not copy over from the formatting on word. will submit abdominal venous once these are committed